scripts and shift notebook
