+++ src/analysis/feature_decomposition.py
@@ -170,26 +170,3 @@ def decompose_activations(
         comp_activ = 1 / (1 + model.transform(mat))
 
     return components, comp_activ, model
-
-
-def project_test_sample(
-    sample: torch.Tensor, analysis_model: Callable, decomposition_type: str = "nndl"
-) -> np.ndarray:
-    """
-    Input:
-        sample: torch tensor or numpy array object of shape (N_samples, Representation_dim). Should contain test representations
-        analysis_model: Already learnt sklearn dictionary learning / clustering object.
-        decomposition_type: Dictionary learning model type (Options: PCA, KMeans, Semi-NMF/Non-negative dict learning, Simple)
-    Output:
-        proj: numpy array of shape (N_samples, # components of analysis_model)
-    """
-    if isinstance(sample, torch.Tensor):
-        sample = sample.cpu().numpy()
-
-    assert isinstance(sample, np.ndarray), "sample should be of type np.ndarray"
-
-    projected_sample = analysis_model.transform(sample)
-    if decomposition_type in ["kmeans", "simple"]:
-        # Kmeans transforms to cluster distances and not "activations". 1/(1+x) transformation to view distances as activations
-        projected_sample = 1 / (1 + projected_sample)
-    return projected_sample
