+++ src/analysis/multimodal_grounding.py
@@ -114,14 +114,14 @@ def get_multimodal_grounding(
         grounding_dict["text_grounding"] = grounded_words
 
     if image_grounding:
-        logger.info("Activations size: ", activations.shape)
+        logger.info(f"Activations size: {activations.shape}")
 
         image_indices = concept_image_grounding(
             activations=activations,
             num_images_per_concept=num_most_activating_samples,
         )
         image_paths = metadata.get("image", [])
-        logger.info("Image paths length: ", len(image_paths))
+        logger.info(f"Image paths length: {len(image_paths)}")
         # Only keep image paths for samples with token_of_interest_mask True
 
         token_of_interest_mask = metadata.get("token_of_interest_mask", None)
