+++ src/metrics/dictionary_learning_metrics.py
@@ -6,7 +6,6 @@ import numpy as np
 import torch
 from nltk.corpus import words
 
-from analysis.feature_decomposition import project_test_samples
 from metrics.clipscore import extract_image_features, img_clipscore
 from metrics.utils import get_stopwords, valid_word
 
@@ -15,7 +14,32 @@ __all__ = [
     "get_random_words",
     "compute_grounding_words_overlap",
     "compute_test_clipscore",
+    "project_test_sample"
 ]
+
+
+def project_test_sample(
+    sample: torch.Tensor, analysis_model: Callable, decomposition_type: str = "nndl"
+) -> np.ndarray:
+    """
+    Input:
+        sample: torch tensor or numpy array object of shape (N_samples, Representation_dim). Should contain test representations
+        analysis_model: Already learnt sklearn dictionary learning / clustering object.
+        decomposition_type: Dictionary learning model type (Options: PCA, KMeans, Semi-NMF/Non-negative dict learning, Simple)
+    Output:
+        proj: numpy array of shape (N_samples, # components of analysis_model)
+    """
+    if isinstance(sample, torch.Tensor):
+        sample = sample.cpu().numpy()
+
+    assert isinstance(sample, np.ndarray), "sample should be of type np.ndarray"
+
+    projected_sample = analysis_model.transform(sample)
+    if decomposition_type in ["kmeans", "simple"]:
+        # Kmeans transforms to cluster distances and not "activations". 1/(1+x) transformation to view distances as activations
+        projected_sample = 1 / (1 + projected_sample)
+    return projected_sample
+
 
 
 def get_clip_score(
@@ -32,7 +56,7 @@ def get_clip_score(
     metadata = list(metadata.values())[0]
     analysis_model = concepts_dict["analysis_model"]
     grounding_words = concepts_dict["text_grounding"]
-    projections = project_test_samples(
+    projections = project_test_sample(
         sample=features,
         analysis_model=analysis_model,
         decomposition_type=concepts_dict["decomposition_method"],
